fix https://github.com/galaxyproject/cargo-port/issues/165
and remove a bit of code duplication
